Export all piece of header and assemble them in wp-story-editor
