+++ packages/story-editor/src/components/header/buttons/index.js
@@ -14,91 +14,7 @@
  * limitations under the License.
  */
 
-/**
- * External dependencies
- */
-import styled from 'styled-components';
-
-/**
- * Internal dependencies
- */
-import { useStory } from '../../../app';
-import CircularProgress from '../../circularProgress';
-import Preview from './preview';
-import SwitchToDraft from './switchToDraft';
-import Update from './update';
-import Publish from './publish';
-
-const ButtonList = styled.nav`
-  display: flex;
-  justify-content: flex-end;
-  padding: 1em;
-  height: 100%;
-`;
-
-const List = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const Space = styled.div`
-  width: 8px;
-`;
-
-const Spinner = styled.div`
-  position: absolute;
-  top: 0;
-`;
-
-const IconWithSpinner = styled.div`
-  position: relative;
-`;
-
-function Loading() {
-  const { isSaving } = useStory((state) => ({
-    isSaving: state.state.meta.isSaving,
-  }));
-  return (
-    isSaving && (
-      <Spinner>
-        <CircularProgress size={32} />
-      </Spinner>
-    )
-  );
-}
-
-function Buttons() {
-  const { status } = useStory(
-    ({
-      state: {
-        story: { status, embedPostLink, link },
-        meta: { isFreshlyPublished },
-      },
-    }) => ({
-      status,
-      embedPostLink,
-      link,
-      isFreshlyPublished,
-    })
-  );
-
-  const isDraft = 'draft' === status;
-
-  return (
-    <ButtonList>
-      <List>
-        <IconWithSpinner>
-          <Preview />
-          <Loading />
-        </IconWithSpinner>
-        <Space />
-        {isDraft ? <Update /> : <SwitchToDraft />}
-        <Space />
-        {isDraft && <Publish />}
-        {!isDraft && <Update />}
-        <Space />
-      </List>
-    </ButtonList>
-  );
-}
-export default Buttons;
+export { default as PreviewButton } from './preview';
+export { default as PublishButton } from './publish';
+export { default as SwitchToDraftButton } from './switchToDraft';
+export { default as UpdateButton } from './update';
