+++ packages/wp-story-editor/package.json
@@ -17,7 +17,8 @@
         "@web-stories-wp/react": "*",
         "@web-stories-wp/story-editor": "*",
         "@web-stories-wp/tracking": "*",
-        "flagged": "^2.0.1"
+        "flagged": "^2.0.1",
+	    "styled-components": "^5.3.1"
     },
     "devDependencies": {
         "@testing-library/react": "^12.0.0"
